move script
